refactor type
